HYP-38: Move remaining api methods out of views.py
Also parameterizes the DEBUG setting and separates out the Manage screen's sections with headings.
