+++ app/projects/views.py
@@ -54,87 +54,6 @@ def signout(request):
 
 
 @user_auth_and_jwt
-def save_dynamic_signed_agreement_form(request):
-    user = request.user
-    agreement_form_id = request.POST['agreement_form_id']
-    project_key = request.POST['project_key']
-    model_name = request.POST['model_name']
-    agreement_text = request.POST['agreement_text']
-
-    save_dynamic_form(agreement_form_id, project_key, model_name, request.POST, user, agreement_text=agreement_text)
-
-    return HttpResponse(200)
-
-
-@user_auth_and_jwt
-def save_signed_agreement_form(request):
-
-    agreement_form_id = request.POST['agreement_form_id']
-    project_key = request.POST['project_key']
-    agreement_text = request.POST['agreement_text']
-
-    agreement_form = AgreementForm.objects.get(id=agreement_form_id)
-    project = DataProject.objects.get(project_key=project_key)
-
-    signed_agreement_form = SignedAgreementForm(user=request.user,
-                                                agreement_form=agreement_form,
-                                                project=project,
-                                                date_signed=datetime.now(),
-                                                agreement_text=agreement_text)
-    signed_agreement_form.save()
-
-    return HttpResponse(200)
-
-
-@user_auth_and_jwt
-def save_signed_external_agreement_form(request):
-    """
-    We cannot track if someone has signed a form on an external website, but we can at least
-    track that they have clicked the link to visit that website. With this record created,
-    an administrator can then manually verify the form on that external site and track their
-    approval within Hypatio.
-    """
-
-    agreement_form_id = request.POST['agreement_form_id']
-    project_key = request.POST['project_key']
-
-    agreement_form = AgreementForm.objects.get(id=agreement_form_id)
-    project = DataProject.objects.get(project_key=project_key)
-
-    # Only create a new record if one does not already exist
-    try:
-        signed_form = SignedAgreementForm.objects.get(
-            user=request.user,
-            agreement_form=agreement_form,
-            project=project
-        )
-    except ObjectDoesNotExist:
-        agreement_text = 'The Participant accessed this form via the 3rd party website. Check there if signed appropriately.'
-        signed_agreement_form = SignedAgreementForm(
-            user=request.user,
-            agreement_form=agreement_form,
-            project=project,
-            date_signed=datetime.now(),
-            agreement_text=agreement_text
-        )
-        signed_agreement_form.save()
-
-    return HttpResponse(200)
-
-
-@user_auth_and_jwt
-def submit_user_permission_request(request):
-    """
-    TODO comment
-    """
-
-    # TODO Update this to create a new permission request record
-    # ...
-
-    return HttpResponse(200)
-
-
-@user_auth_and_jwt
 def signed_agreement_form(request):
 
     project_key = request.GET['project_key']
