+++ app/templates/manage/base.html
@@ -16,8 +16,14 @@
 
 {% block content %}
 <div class="row">
+    <div class="col-md-12" style="margin-bottom: 15px;">
+        <h3>Essentials</h3>
+    </div>
+</div>
+
+<div class="row">
     <div class="col-md-3">
-        <div class="panel panel-primary">
+        <div class="panel panel-default">
             <div class="panel-heading" role="tab" id="heading-description">
                 <h2 class="panel-title">Registration</h2>
             </div>
@@ -34,7 +40,7 @@
     </div>
 
     <div class="col-md-9">
-        <div class="panel panel-primary">
+        <div class="panel panel-default">
             <div class="panel-heading" role="tab" id="heading-description">
                 <h2 class="panel-title">Details</h2>
             </div>
@@ -45,7 +51,7 @@
     </div>
 
     <!-- <div class="col-md-4">
-        <div class="panel panel-primary">
+        <div class="panel panel-default">
             <div class="panel-heading" role="tab" id="heading-description">
                 <h2 class="panel-title">Reports and Statistics</h2>
             </div>
@@ -57,8 +63,15 @@
 </div>
 
 <div class="row">
+    <div class="col-md-12" style="margin-bottom: 15px;">
+        <hr>
+        <h3>Forms, Files, and Tasks</h3>
+    </div>
+</div>
+
+<div class="row">
     <div class="col-lg-4">
-        <div class="panel panel-primary">
+        <div class="panel panel-default">
             <div class="panel-heading" role="tab" id="heading-description">
                 <h2 class="panel-title">Agreement Forms</h2>
             </div>
@@ -90,7 +103,7 @@
     </div>
 
     <div class="col-lg-4">
-        <div class="panel panel-primary">
+        <div class="panel panel-default">
             <div class="panel-heading" role="tab" id="heading-description">
                 <h2 class="panel-title">Files</h2>
             </div>
@@ -133,7 +146,7 @@
     </div>
 
     <div class="col-lg-4">
-        <div class="panel panel-primary">
+        <div class="panel panel-default">
             <div class="panel-heading" role="tab" id="heading-description">
                 <h2 class="panel-title">Tasks</h2>
             </div>
@@ -145,8 +158,15 @@
 </div>
 
 <div class="row">
+    <div class="col-md-12" style="margin-bottom: 15px;">
+        <hr>
+        <h3>Reports and Visualizations</h3>
+    </div>
+</div>
+
+<div class="row">
     <div class="col-lg-4">
-        <div class="panel panel-primary">
+        <div class="panel panel-default">
             <div class="panel-heading" role="tab" id="heading-description">
                 <h2 class="panel-title">Reports</h2>
             </div>
@@ -162,11 +182,8 @@
             </div>
         </div>
     </div>
-</div>
-
-<div class="row">
-    <div class="col-md-6">
-        <div class="panel panel-primary">
+    <div class="col-md-4">
+        <div class="panel panel-default">
             <div class="panel-heading" role="tab" id="heading-description">
                 <h2 class="panel-title">Downloads</h2>
             </div>
@@ -175,8 +192,8 @@
             </div>
         </div>
     </div>
-    <div class="col-md-6">
-        <div class="panel panel-primary">
+    <div class="col-md-4">
+        <div class="panel panel-default">
             <div class="panel-heading" role="tab" id="heading-description">
                 <h2 class="panel-title">Uploads</h2>
             </div>
@@ -188,9 +205,16 @@
 </div>
 
 <div class="row">
+    <div class="col-md-12" style="margin-bottom: 15px;">
+        <hr>
+        <h3>Team and Participant Management</h3>
+    </div>
+</div>
+
+<div class="row">
     <div class="col-lg-12">
         {% if project.has_teams %}
-        <div class="panel panel-primary">
+        <div class="panel panel-default">
             <div class="panel-heading" role="tab" id="heading-description">
                 <h2 class="panel-title">Team List</h2>
             </div>
@@ -237,7 +261,7 @@
 
 <div class="row">
     <div class="col-lg-12">
-        <div class="panel panel-primary">
+        <div class="panel panel-default">
             <div class="panel-heading" role="tab" id="heading-description">
                 <h2 class="panel-title">Full Participant List</h2>
             </div>
